+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 - Lights are now completely calculated with CSS variables
+- Simplified themes
 ### Fixed
 - [upstream] QR Codes are now responsive
 - The wheel display is mobile friendly now